Database helper method corrections

+++ db_handler.py
@@ -14,22 +14,45 @@ class DBHandler:
         #user = {"name":name,"IP":IP,"port":port }
 
         try:
-            user = {"_id":name,"IP":ip,"port":port }
-            self.user_collection.insert_one(user)
-            return True 
+            
+            if (name):
+            
+                if (self.user_collection.count_documents({"_id":name}) == 0):
+            
+                    user = {"_id":name,"IP":ip,"port":port }
+                    self.user_collection.insert_one(user)
+                    return 0
+            
+                else:
+
+                    print("USERNAME:" + str(name) + " ALREADY EXISTS!")
+                    return 1
+
+            else:
+                
+                print("NOT A VALID USERNAME!")
+                return 2
 
         except:
 
-            print("USERNAME:" + str(name) + " ALREADY EXISTS!")
-            return False
+            print("USERNAME NOT ADDED! DATABASE CONNECTION ERROR")
+            return 3
 
     #Assuming unique name IDs for each user
     def remove_user(self,name):
 
         try:
-            user_query = {"_id":name}
-            self.user_collection.delete_one(user_query)
-            return True 
+
+            if (self.user_collection.count_documents({"_id":name}) != 0):
+                
+                user_query = {"_id":name}
+                self.user_collection.delete_one(user_query)
+                return True 
+            
+            else:
+
+                print("USERNAME:" + str(name) + " DOES NOT EXIST!")
+                return False
 
         except:
 
@@ -39,10 +62,18 @@ class DBHandler:
     def update_socket(self,name,ip,port):
 
         try:
-            user_query = {"_id":name}
-            user_newvalues = { "$set": {"IP": ip,"port": port}}
-            self.user_collection.update_one(user_query, user_newvalues)
-            return True 
+
+            if (self.user_collection.count_documents({"_id":name}) != 0):
+
+                user_query = {"_id":name}
+                user_newvalues = { "$set": {"IP": ip,"port": port}}
+                self.user_collection.update_one(user_query, user_newvalues)
+                return True 
+            
+            else:
+
+                print("USERNAME:" + str(name) + " DOES NOT EXIST!")
+                return False
         
         except:
             print("USERNAME:" + str(name) + "DOES NOT EXIST!")
@@ -53,11 +84,18 @@ class DBHandler:
 
         try: 
 
-            user_query = {"_id":name}
-            user_newvalues = {"$set": {"subjects": subjects}}
-            self.user_collection.update_one(user_query,user_newvalues)
+            if (self.user_collection.count_documents({"_id":name}) != 0):
+
+                user_query = {"_id":name}
+                user_newvalues = {"$set": {"subjects": subjects}}
+                self.user_collection.update_one(user_query,user_newvalues)
+                return True 
             
-            return True 
+            else:
+
+                print("USERNAME:" + str(name) + " DOES NOT EXIST!")
+                return False
+
         
         except:
             print("USERNAME:" + str(name) + "DOES NOT EXIST!")
@@ -77,7 +115,7 @@ class DBHandler:
 
                     print("subject exits")
                 
-                    message = {"name":name,"subject":subject,"text":text}
+                    message = {"name":name,"subject":subject,"text":text,"clients_received":[]}
                     self.messages_collection.insert_one(message)
                     return True
 
@@ -98,28 +136,35 @@ class DBHandler:
     def retrieve_texts (self, name):
 
         try:
-            
-            user_cursor = self.user_collection.find_one({"_id":{"$eq":name}})
 
-            user_subjects = user_cursor["subjects"]
+            if (self.user_collection.count_documents({"_id":name}) != 0):
+            
+                user_cursor = self.user_collection.find_one({"_id":{"$eq":name}})
 
-            #for subject in user_subjects:
-            #    message = self.messages_collection.find_one({"subject":{"$eq":subject}})
-            #   print(message["text"])
-            
-            #print(user_subjects[0])
-            
-            message_cursor = self.messages_collection.find({"subject":{"$in":user_subjects}})
-            #print(message_cursor["text"])
-            
-            msg_list = []
-            
-            for document in message_cursor:
+                user_subjects = user_cursor["subjects"]
+        
+
+                message_query = {"$and":[{"subject":{"$in":user_subjects}},{"clients_received":{"$nin":[name]}}]}
+                message_cursor = self.messages_collection.find(message_query)
                 
-                msg = {"NAME":document["name"],"SUBJECT":document["subject"],"TEXT":document["text"]}
-                msg_list.append(msg)  
+                message_newClients = {"$push": {"clients_received":name}}
+                
+                msg_list = []
+
+                
+                for document in message_cursor:
+
+                    msg = {"NAME":document["name"],"SUBJECT":document["subject"],"TEXT":document["text"]}
+                    msg_list.append(msg)  
             
-            return msg_list
+                self.messages_collection.update_many(message_query,message_newClients)
+                
+                return msg_list
+            
+            else:
+
+                print("USERNAME:" + str(name) + " DOES NOT EXIST!")
+                return 
         
         except:
